requirements.txt
flask
flask-cors
requests
openmeteo-requests
requests-cache
retry-requests
numpy
pandas
python-dotenv
